NIFI-6016 PutCassandraRecord batch size
Adds resetting the batch size to fix broken batch processing

Removes empty line


This closes #3337
