+++ nifi-nar-bundles/nifi-druid-bundle/nifi-druid-controller-service/src/main/java/org/apache/nifi/controller/druid/DruidTranquilityController.java
@@ -80,7 +80,6 @@ public class DruidTranquilityController extends AbstractControllerService implem
     private final static String FIREHOSE_PATTERN = "druid:firehose:%s";
 
     private final static AllowableValue PT1M = new AllowableValue("PT1M", "1 minute", "1 minute");
-    private final static AllowableValue PT5M = new AllowableValue("PT5M", "5 minutes", "5 minutes");
     private final static AllowableValue PT10M = new AllowableValue("PT10M", "10 minutes", "10 minutes");
     private final static AllowableValue PT60M = new AllowableValue("PT60M", "60 minutes", "1 hour");
 
@@ -260,7 +259,7 @@ public class DruidTranquilityController extends AbstractControllerService implem
     public static final PropertyDescriptor INDEX_RETRY_PERIOD = new PropertyDescriptor.Builder()
             .name("druid-cs-index-retry-period")
             .displayName("Index Retry Period")
-            .description("Time period, until a transiently failing indexing service overlord call is retried, before giving up.")
+            .description("Grace period to allow late arriving events for real time ingest.")
             .required(true)
             .allowableValues(PT1M, PT10M, PT60M)
             .defaultValue(PT10M.getValue())
@@ -274,16 +273,6 @@ public class DruidTranquilityController extends AbstractControllerService implem
             .required(true)
             .allowableValues(PT1M, PT10M, PT60M)
             .defaultValue(PT10M.getValue())
-            .addValidator(StandardValidators.NON_EMPTY_VALIDATOR)
-            .build();
-
-    public static final PropertyDescriptor FIREHOSE_GRACE_PERIOD = new PropertyDescriptor.Builder()
-            .name("druid-cs-firehose-grace-period")
-            .displayName("Firehose Grace Period")
-            .description("An additional grace period, after the \"Late Event Grace Period\" (window period) has elapsed, but before the indexing task is shut down.")
-            .required(true)
-            .allowableValues(PT1M, PT5M, PT10M, PT60M)
-            .defaultValue(PT5M.getValue())
             .addValidator(StandardValidators.NON_EMPTY_VALIDATOR)
             .build();
 
@@ -341,9 +330,7 @@ public class DruidTranquilityController extends AbstractControllerService implem
         props.add(AGGREGATOR_JSON);
         props.add(SEGMENT_GRANULARITY);
         props.add(QUERY_GRANULARITY);
-        props.add(INDEX_RETRY_PERIOD);
         props.add(WINDOW_PERIOD);
-        props.add(FIREHOSE_GRACE_PERIOD);
         props.add(TIMESTAMP_FIELD);
         props.add(MAX_BATCH_SIZE);
         props.add(MAX_PENDING_BATCHES);
@@ -390,7 +377,6 @@ public class DruidTranquilityController extends AbstractControllerService implem
         final String segmentGranularity = context.getProperty(SEGMENT_GRANULARITY).getValue();
         final String queryGranularity = context.getProperty(QUERY_GRANULARITY).getValue();
         final String windowPeriod = context.getProperty(WINDOW_PERIOD).getValue();
-        final String firehoseGracePeriod = context.getProperty(FIREHOSE_GRACE_PERIOD).getValue();
         final String indexRetryPeriod = context.getProperty(INDEX_RETRY_PERIOD).getValue();
         final String aggregatorJSON = context.getProperty(AGGREGATOR_JSON).evaluateAttributeExpressions().getValue();
         final String dimensionsStringList = context.getProperty(DIMENSIONS_LIST).evaluateAttributeExpressions().getValue();
@@ -429,7 +415,7 @@ public class DruidTranquilityController extends AbstractControllerService implem
         final TimestampSpec timestampSpec = new TimestampSpec(timestampField, "auto", null);
 
         final Beam<Map<String, Object>> beam = buildBeam(dataSource, indexService, discoveryPath, clusterPartitions, clusterReplication,
-                segmentGranularity, queryGranularity, windowPeriod, firehoseGracePeriod, indexRetryPeriod, dimensions, aggregator, timestamper, timestampSpec);
+                segmentGranularity, queryGranularity, windowPeriod, indexRetryPeriod, dimensions, aggregator, timestamper, timestampSpec);
 
         tranquilizer = buildTranquilizer(maxBatchSize, maxPendingBatches, lingerMillis, beam);
 
@@ -446,7 +432,7 @@ public class DruidTranquilityController extends AbstractControllerService implem
     }
 
     Beam<Map<String, Object>> buildBeam(String dataSource, String indexService, String discoveryPath, int clusterPartitions, int clusterReplication,
-                                        String segmentGranularity, String queryGranularity, String windowPeriod, String firehoseGracePeriod, String indexRetryPeriod, List<String> dimensions,
+                                        String segmentGranularity, String queryGranularity, String windowPeriod, String indexRetryPeriod, List<String> dimensions,
                                         List<AggregatorFactory> aggregator, Timestamper<Map<String, Object>> timestamper, TimestampSpec timestampSpec) {
         return DruidBeams.builder(timestamper)
                 .curator(curator)
@@ -467,7 +453,6 @@ public class DruidTranquilityController extends AbstractControllerService implem
                         DruidBeamConfig
                                 .builder()
                                 .indexRetryPeriod(new Period(indexRetryPeriod))
-                                .firehoseGracePeriod(new Period(firehoseGracePeriod))
                                 .build())
                 .buildBeam();
     }
