+++ nifi-nar-bundles/nifi-framework-bundle/nifi-framework/nifi-web/nifi-web-ui/src/main/webapp/js/nf/canvas/nf-canvas.js
@@ -651,9 +651,6 @@
                 });
             });
 
-            // don't let the reload action get called more than once every second
-            var throttledCanvasReload = nfCommon.throttle(nfActions.reload, 1000);
-
             // listen for browser resize events to reset the graph size
             $(window).on('resize', function (e) {
                 if (e.target === window) {
@@ -711,7 +708,7 @@
                         nfCommon.toggleScrollable(tabsContent.get(0));
                     });
                 }
-            }).on('keydown', function (evt) {
+            }).on('keyup', function (evt) {
                 // if a dialog is open, disable canvas shortcuts
                 if ($('.dialog').is(':visible') || $('#search-field').is(':focus')) {
                     return;
@@ -729,7 +726,7 @@
                             return;
                         }
                         // ctrl-r
-                        throttledCanvasReload();
+                        nfActions.reload();
 
                         // default prevented in nf-universal-capture.js
                     } else if (evt.keyCode === 65) {
