+++ nifi-nar-bundles/nifi-scripting-bundle/nifi-scripting-processors/src/main/java/org/apache/nifi/script/impl/JythonScriptEngineConfigurator.java
@@ -18,7 +18,6 @@ package org.apache.nifi.script.impl;
 
 import org.apache.nifi.logging.ComponentLog;
 import org.apache.nifi.processors.script.ScriptEngineConfigurator;
-import org.python.core.PyString;
 
 import javax.script.ScriptEngine;
 import javax.script.ScriptException;
@@ -48,7 +47,7 @@ public class JythonScriptEngineConfigurator implements ScriptEngineConfigurator 
             engine.eval("import sys");
             if (modulePaths != null) {
                 for (String modulePath : modulePaths) {
-                    engine.eval("sys.path.append(" + PyString.encode_UnicodeEscape(modulePath, true) + ")");
+                    engine.eval("sys.path.append('" + modulePath + "')");
                 }
             }
         }
