+++ nifi-nar-bundles/nifi-scripting-bundle/nifi-scripting-processors/src/main/resources/META-INF/services/org.apache.nifi.controller.ControllerService
@@ -16,4 +16,3 @@
 org.apache.nifi.record.script.ScriptedReader
 org.apache.nifi.record.script.ScriptedRecordSetWriter
 org.apache.nifi.lookup.script.ScriptedLookupService
-org.apache.nifi.lookup.script.SimpleScriptedLookupService
