+++ nifi-nar-bundles/nifi-scripting-bundle/nifi-scripting-processors/src/test/java/org/apache/nifi/processors/script/TestInvokeJython.java
@@ -16,7 +16,6 @@
  */
 package org.apache.nifi.processors.script;
 
-import org.apache.commons.lang3.StringUtils;
 import org.apache.nifi.components.ValidationResult;
 import org.apache.nifi.script.ScriptingComponentUtils;
 import org.apache.nifi.util.MockFlowFile;
@@ -32,9 +31,6 @@ import java.util.Collection;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.Arrays;
-import java.util.stream.Collectors;
-import java.util.stream.IntStream;
 
 public class TestInvokeJython extends BaseScriptTest {
 
@@ -96,57 +92,6 @@ public class TestInvokeJython extends BaseScriptTest {
 
         // verify reading an attribute value
         result.get(0).assertAttributeEquals("from-attribute", "value-2");
-    }
-
-    /**
-     * Test a script that has a Jython processor that reads the system path as controlled by the Module Directory property then stores it in the attributes of the flowfile being routed.
-     * <p>
-     * This tests whether the JythonScriptEngineConfigurator successfully translates the "Module Directory"  property into Python system paths, even with strings that contain Python escape sequences
-     *
-     * @throws Exception Any error encountered while testing
-     */
-    @Test
-    public void testUpdateAttributeFromProcessorModulePaths() throws Exception {
-        // Prepare a set of easily identified paths for the Module Directory property
-        final String moduleDirectoryTestPrefix = "test";
-        final String[] testModuleDirectoryValues = { "abc","\\a\\b\\c","\\123","\\d\"e" };
-        final int numTestValues = testModuleDirectoryValues.length;
-        // Prepend each module directory value with a simple prefix and an identifying number so we can identify it later.
-        final List<String> testModuleDirectoryFullValues = IntStream.range(0,numTestValues)
-                .boxed()
-                .map(i -> String.format("%s#%s#%s",moduleDirectoryTestPrefix,i,testModuleDirectoryValues[i]))
-                .collect(Collectors.toList());
-        final String testModuleDirectoryCombined = String.join(",",testModuleDirectoryFullValues);
-
-        // Run the script that captures the system path resulting from the Module Directory property
-        final TestRunner runner = TestRunners.newTestRunner(new InvokeScriptedProcessor());
-
-        runner.setValidateExpressionUsage(false);
-        runner.setProperty(scriptingComponent.getScriptingComponentHelper().SCRIPT_ENGINE, "python");
-        runner.setProperty(ScriptingComponentUtils.SCRIPT_FILE, "target/test/resources/jython/test_modules_path.py");
-        runner.setProperty(ScriptingComponentUtils.MODULES, testModuleDirectoryCombined);
-
-        final Map<String, String> attributes = new HashMap<>();
-
-        runner.assertValid();
-        runner.enqueue(new byte[0], attributes);
-        runner.run();
-
-        runner.assertAllFlowFilesTransferred("success", 1);
-        final List<MockFlowFile> result = runner.getFlowFilesForRelationship("success");
-
-        // verify successful processing of the module paths
-        result.get(0).assertAttributeExists("from-path");
-        final String[] effectivePaths = result.get(0).getAttribute("from-path").split(","); // Extract the comma-delimited paths from the script-produced attribute
-        Assert.assertTrue(effectivePaths.length >= numTestValues); // we should have our test values, plus defaults
-        // Isolate only the paths with our identified prefix
-        final List<String> relevantPaths = Arrays.stream(effectivePaths).filter(path -> path.startsWith(moduleDirectoryTestPrefix)).collect(Collectors.toList());
-        Assert.assertEquals(testModuleDirectoryFullValues.size(), relevantPaths.size());
-        relevantPaths.forEach(path -> {
-            final int resultIx = Integer.valueOf(StringUtils.substringBetween(path,"#")); // extract the index so we can relate it to the sources, despite potential mangling
-            final String expectedValue = testModuleDirectoryFullValues.get(resultIx);
-            Assert.assertEquals(expectedValue, path); // Ensure our path was passed through without mangling
-        });
     }
 
     /**
